// src/domain/models/type/user-activity.ts
export type SessionsApiActivity = {
    day: string;
    kilogram: number;
    calories: number;
};

export type UserActivityApi = {
    userId: number;
    sessions: SessionsApiActivity[];
};

export type SessionsActivity = {
    day: string;
    kilogram: number;
    calories: number;
};

export type UserActivity = {
    userId: number;
    sessions: SessionsActivity[];
};